migrate tpyes for usercontroller/authManagement to types folder
